Working Version
The GUI is very simple but it works.
Test code is included but might not work if you do not supply images. To
make it run, create a test_code folder inside Solstice. Place images
inside that folder with the same name as that inside the test code or
just change the path of the image inside the test code.