Took default args from WorldConverter constructor
In case you keep changing the tile size of the game, this will be handy
since you will be forced to specify a size and change it later. The best
practice will be to have constants for the tile size and create an
instance as: WorldGridConverter(TILE_WIDTH, TILE_HEIGHT).